+++ policy/src/test/java/org/apache/brooklyn/policy/loadbalancing/LoadBalancingPolicyConcurrencyTest.java
@@ -191,7 +191,7 @@ public class LoadBalancingPolicyConcurrencyTest extends AbstractLoadBalancingPol
                 }});
         }
         
-        assertWorkratesEventually(containers, items, Collections.nCopies((int)(NUM_CONTAINERS/2), 40d), WORKRATE_JITTER*2);
+        assertWorkratesEventually(containers, items, Collections.nCopies(NUM_CONTAINERS/2, 40d), WORKRATE_JITTER*2);
     }
     
     @Test(groups="WIP")
